+++ dart/dynamics/EllipsoidShape.hpp
@@ -37,7 +37,8 @@
 namespace dart {
 namespace dynamics {
 
-class EllipsoidShape : public Shape {
+class EllipsoidShape : public Shape
+{
 public:
   /// \brief Constructor.
   explicit EllipsoidShape(const Eigen::Vector3d& diameters);
@@ -54,14 +55,20 @@ public:
   static const std::string& getStaticType();
 
   /// \brief Set diameters of this ellipsoid.
-  /// \deprecated Deprecated in 6.2. Please use setRadii() instead.
+  /// \deprecated Deprecated in 6.2. Please use setDiameters() instead.
   DART_DEPRECATED(6.2)
-  void setSize(const Eigen::Vector3d& _size);
+  void setSize(const Eigen::Vector3d& diameters);
 
   /// \brief Get diameters of this ellipsoid.
-  /// \deprecated Deprecated in 6.2. Please use getRadii() instead.
+  /// \deprecated Deprecated in 6.2. Please use getDiameters() instead.
   DART_DEPRECATED(6.2)
   const Eigen::Vector3d& getSize() const;
+
+  /// \brief Set diameters of this ellipsoid.
+  void setDiameters(const Eigen::Vector3d& diameters);
+
+  /// \brief Get diameters of this ellipsoid.
+  const Eigen::Vector3d& getDiameters() const;
 
   /// Set radii of this ellipsoid.
   void setRadii(const Eigen::Vector3d& radii);
@@ -70,16 +77,20 @@ public:
   const Eigen::Vector3d getRadii() const;
 
   /// \brief Compute volume from given properties
-  static double computeVolume(const Eigen::Vector3d& size);
+  static double computeVolume(const Eigen::Vector3d& diameters);
+  // TODO(JS): In order to follow the commonly used convention, change to take
+  // radii instead of diameters in DART 7.
 
   /// \brief Compute moments of inertia of a ellipsoid
   static Eigen::Matrix3d computeInertia(
-      const Eigen::Vector3d& size, double mass);
+      const Eigen::Vector3d& diameters, double mass);
+  // TODO(JS): In order to follow the commonly used convention, change to take
+  // radii instead of diameters in DART 7.
 
   // Documentation inherited.
   Eigen::Matrix3d computeInertia(double mass) const override;
 
-  /// \brief True if (mDim[0] == mDim[1] == mDim[2]).
+  /// \brief True if all the radii are exactly eqaul.
   bool isSphere(void) const;
 
 protected:
@@ -87,8 +98,8 @@ protected:
   void updateVolume() override;
 
 private:
-  /// \brief Size of this ellipsoid
-  Eigen::Vector3d mSize;
+  /// \brief Diameters of this ellipsoid
+  Eigen::Vector3d mDiameters;
 
 public:
   // To get byte-aligned Eigen vectors