+++ dart/gui/OpenGLRenderInterface.hpp
@@ -81,7 +81,7 @@ public:
     GLuint compileList(const Eigen::Vector3d& _scale, const aiScene* _mesh);
 
     void drawSphere(double _radius) override;
-    void drawEllipsoid(const Eigen::Vector3d& _size) override;
+    void drawEllipsoid(const Eigen::Vector3d& _diameters) override;
     void drawCube(const Eigen::Vector3d& _size) override;
     void drawCylinder(double _radius, double _height) override;
     void drawCapsule(double radius, double height) override;